Split activation.rs into focused sub-modules (#607)
Split src/analysis/activation.rs (1,130 lines) into:
- activation/functions.rs — CPU activation function implementations
- activation/specs.rs — candidate specs, GPU ID mapping, bias helpers
- activation/simulation.rs — target simulation, predicates, variance checking
- activation/mod.rs — public API re-exports and tests

Public API unchanged. All existing tests pass without modification.
